Added Files
These are stuffs I've worked on relatively recently, however, the RAT stuff is just the thing I am most proud of that I have done in python. The RAT client is for a remote administration tool, it is what the user has, the RATserveroutput is for receiving messages from the RATclient and the RATserverinput is for inputting commands to the client. The csgoDLL is for an internal CSGO cheat for csgo that has a triggerbot, bhop, and no-flash. I don't know if the offsets are up to date or not so it may or may not still work but all I would need to make it work would be to update the offsets. I also made an esp but I followed the youtube tutorial a little more than I should have to where I'm not going to post it as my own. These are by no means my only projects, I have also done web scraping with python and bots using OpenCV that screenshotted, checked the pixel color and then jumped in the chrome dinosaur game. (That is just a project that came to mind I have done much more than just that.)